==> Scripts/Script03_mind-wandering_prediction_webgazerdata.ipynb <==
In this document, we first extract learners' mind-wandering from the json data. Then we map the data on the timeline. We process the mind-wandering data generated by reporting and questions seperately.

## 1. Data Reading
In this step, we read gaze data of Tobii from a tsv file and mind-wandering reports from a json file.

import pandas as pd
import numpy as np
from pandas import DataFrame

# folderpath_webgazerdata_rprocessed = "../Data_Publish/Data_WebGazer/Data_ProcessedByScript02"
folderpath_webgazerdata_rprocessed = "../Data_Publish/Data_Tobii/Data_ProcessedByScript02"
folderpath_webgazerdata = "../Data_Publish/Data_Event"

### 1.1 Reading Mind-Wandering Data
We only read some important data columns from the json file.

1. Rating
2. Bell Rings

import json
import os

list_filepath_report = []
list_id_report = []
for file in os.listdir(folderpath_webgazerdata):
    if file.endswith(".json"):
        list_filepath_report.append(os.path.join(folderpath_webgazerdata, file))
        list_id_report.append(file[0:-5])
        
print list_filepath_report
print list_id_report

#### Data Format 
id, starttime_iso, endtime_iso, starttime_video, endtime_video, video_length, label

import datetime

df_reports = pd.DataFrame()

for i in range(0, len(list_id_report)):
    # id
    id_str = list_id_report[i]
    print id_str
    # starttime_iso, endtime_iso, starttime_video, endtime_video, video_length, label
    with open(list_filepath_report[i]) as file_json_data:
        json_data = json.load(file_json_data)
        # print(json_data['activity'])
        
        # TODO: full screen playing info
        fullscreen_list = []
        fullscreen_temp = {'enter': '', 'exit': ''}
        for videostatus in json_data['user']['videostatus']:
            if videostatus['status'] == 'Fullscreen_enter':
                fullscreen_temp = {'enter': '', 'exit': ''}
                fullscreen_temp['enter'] = videostatus['time']
            elif videostatus['status'] == 'Fullscreen_exit':
                if (fullscreen_temp['enter'] != ''):
                    fullscreen_temp['exit'] = videostatus['time']
                    fullscreen_list.append(fullscreen_temp)
                    fullscreen_temp = {'enter': '', 'exit': ''}
            elif videostatus['status'] == 'ENDED':
                if (fullscreen_temp['enter'] != ''):
                    fullscreen_temp['exit'] = videostatus['time']
                    fullscreen_list.append(fullscreen_temp)
                    fullscreen_temp = {'enter': '', 'exit': ''}           
        
        print fullscreen_list
        
        pre_video_id = ""
        video_order = 1
        
        for bell in json_data['user']['ratingbells']:
            ## End time is the time when bell rings
            endtime_iso = bell['time']
            # print endtime_iso
            endtime_iso_datetime = datetime.datetime.strptime(endtime_iso, "%Y-%m-%dT%H:%M:%S.%fZ")
            ## Start time is 30 sec before the end time
            starttime_iso_datetime = endtime_iso_datetime - datetime.timedelta(seconds=30)
            (dt, micro) = starttime_iso_datetime.strftime('%Y-%m-%dT%H:%M:%S.%f').split('.')
            starttime_iso = "%s.%03dZ" % (dt, int(micro) / 1000)
            # print starttime_iso
            
            endtime_video = bell['videoTime']
            # print endtime_video
            ## There is no stop in last 30 sec before the bell rings. 
            ## Since each time the video playing starts, they will ring the bell after 30 sec.
            starttime_video = str(float(endtime_video) - 30)
            # print starttime_video
            
            video_length = bell['videoDuration']
            # print video_length

            video_id = ""           
            if float(video_length) < 420:
                video_id = "Nuclear"  
            else:
                video_id = "Solar"
            
            if pre_video_id == "":
                video_order = 1
                pre_video_id = video_id
            elif video_id != pre_video_id:
                video_order = video_order + 1
                pre_video_id = video_id
            
            label = 0
            for rating in json_data['user']['ratings']:
                ratingtime_iso = rating['time']
                # print ratingtime_iso
                ratingtime_iso_datetime = datetime.datetime.strptime(ratingtime_iso, "%Y-%m-%dT%H:%M:%S.%fZ")
                if ratingtime_iso_datetime > endtime_iso_datetime and ratingtime_iso_datetime < endtime_iso_datetime + datetime.timedelta(seconds=10):
                    label = 1
                    break
            # print label          
            
            fullscreen_flag = 0
            for fullscreen_play in fullscreen_list:
                if (starttime_iso > fullscreen_play['enter']) and (endtime_iso < fullscreen_play['exit']):
                    fullscreen_flag = 1
                    break
            
            ## Add data into dataframe
            df_reports = df_reports.append({'id': id_str,
                                            'video_id': video_id,
                                            'video_order': video_order,
                                            'starttime_iso': starttime_iso, 
                                            'endtime_iso': endtime_iso,
                                            'starttime_video': starttime_video,
                                            'endtime_video': endtime_video,
                                            'video_length': video_length,
                                            'label': label,
                                            'fullscreen': fullscreen_flag
                                           }, 
                                           ignore_index=True)
print df_reports.head(10)
print df_reports.shape
print df_reports[df_reports['label'] == 1].shape
print df_reports[df_reports['label'] == 0].shape
print df_reports.video_order.values
print df_reports.video_id.values

### 1.1 Reading WebGazer Gaze Data
We only read some important data columns from the tsv file.

## Feature Extraction

# Data format: id, starttime_iso, endtime_iso, feature 1, feature 2.......
import math
from scipy.stats import kurtosis
from scipy.stats import skew

df_features = pd.DataFrame()

for i in range(0, len(list_id_report)):
    # id
    id_str = list_id_report[i]
    print id_str
    df_reports_withid = df_reports.loc[df_reports['id'] == id_str]
    
    ## Read the tsv file based on id_str
    path_gazedata_rprocessed = os.path.join(folderpath_webgazerdata_rprocessed, id_str + ".csv")
    print path_gazedata_rprocessed
    df_GazeData_rprocessed = DataFrame.from_csv(path_gazedata_rprocessed, sep=",")
    # print df_GazeData_Tobii.head(5)
    df_GazeData_rprocessed = df_GazeData_rprocessed.reset_index()
    
    ## Remove unnessesary data   
    df_GazeData_rprocessed = df_GazeData_rprocessed[['Timestamp_utc',
                                                     'FixationIndex',
                                                     'GazeEventDuration',
                                                     'FixationPointX..MCSpx.',
                                                     'FixationPointY..MCSpx.',
                                                     'AbsoluteSaccadicDirection']]
    df_GazeData_rprocessed.columns = ["Timestamp_utc", 
                                      "FixationIndex", 
                                      "GazeEventDuration", 
                                      "FixationPointX (MCSpx)", 
                                      "FixationPointY (MCSpx)",
                                      "AbsoluteSaccadicDirection"]
    
    # TODO:change the format of Timestamp_utc
    def timestamp_trans(row):
        temp_str = row['Timestamp_utc']
        temp_str = temp_str+"Z"
        temp_str = temp_str.replace(" ", "T")
        return temp_str
    
    df_GazeData_rprocessed['Timestamp_utc'] = df_GazeData_rprocessed.apply(timestamp_trans,axis=1)

    df_GazeData_Tobii = df_GazeData_rprocessed
    
    for index, row in df_reports_withid.iterrows():
        starttime_iso = row['starttime_iso']
        print starttime_iso
        endtime_iso = row['endtime_iso']
        print endtime_iso
        
        ## Select Data from df_GazeData_Tobii based on starttime_iso and endtime_iso
        df_GazeData_Tobii_selected = df_GazeData_Tobii.loc[((df_GazeData_Tobii['Timestamp_utc'] >= starttime_iso) &
                                                           (df_GazeData_Tobii['Timestamp_utc'] <= endtime_iso))
                                                          ]
        
        # print df_GazeData_Tobii_selected.head(20)
        # print df_GazeData_Tobii_selected.shape
        # print df_GazeData_Tobii_selected.columns
        
        ## Global Features: Feature Selection based on selected data
        temp_fixationindex = 0
        temp_timestamp = ""
        temp_FixationPointX = 0
        temp_FixationPointY = 0
        list_fixationduration = []
        list_saccadeduration = []
        list_saccadedistance = []
        list_saccadeangel = []
        
        ## Local Features: Feature Selection based on selected data
        
        # TODO: get the info about the video and fullscreen playing
        video_length = row['video_length']
        fullscreen_flag = row['fullscreen']
        
        face_topleft_x = 0
        face_topleft_y = 0
        face_bottomright_x = 0
        face_bottomright_y = 0
                
        slide_topleft_x = 0
        slide_topleft_y = 0
        slide_bottomright_x = 0
        slide_bottomright_y = 0
                
        subtitle_topleft_x = 0
        subtitle_topleft_y = 0
        subtitle_bottomright_x = 0
        subtitle_bottomright_y = 0
        
        # Select coordinate boundaries for the 
        if video_length <= 410: # Nucl should be replaced by video_id
            if fullscreen_flag:
                
                face_topleft_x = 1191
                face_topleft_y = 239
                face_bottomright_x = 1191+331
                face_bottomright_y = 239+280
                
                slide_topleft_x = 137
                slide_topleft_y = 132
                slide_bottomright_x = 137+929
                slide_bottomright_y = 132+557
                
                subtitle_topleft_x = 402
                subtitle_topleft_y = 892
                subtitle_bottomright_x = 402+1134
                subtitle_bottomright_y = 892+134
            
            else:
                
                face_topleft_x = 1089
                face_topleft_y = 297
                face_bottomright_x = 1089+139
                face_bottomright_y = 297+141
                
                slide_topleft_x = 558
                slide_topleft_y = 246
                slide_bottomright_x = 558+456
                slide_bottomright_y = 246+273
                
                subtitle_topleft_x = 721
                subtitle_topleft_y = 617
                subtitle_bottomright_x = 721+478
                subtitle_bottomright_y = 617+67
        
        elif video_length > 450: # Solar
            
            if fullscreen_flag:
                
                face_topleft_x = 501
                face_topleft_y = 189
                face_bottomright_x = 501+267
                face_bottomright_y = 189+260
                
                slide_topleft_x = 861
                slide_topleft_y = 313
                slide_bottomright_x = 861+811
                slide_bottomright_y = 313+483
                
                subtitle_topleft_x = 458
                subtitle_topleft_y = 900
                subtitle_bottomright_x = 458+1010
                subtitle_bottomright_y = 900+128
            
            else:
                
                face_topleft_x = 721
                face_topleft_y = 263
                face_bottomright_x = 721+146
                face_bottomright_y = 263+139
                
                slide_topleft_x = 913
                slide_topleft_y = 330
                slide_bottomright_x = 913+401
                slide_bottomright_y = 330+241
                
                subtitle_topleft_x = 709
                subtitle_topleft_y = 622
                subtitle_bottomright_x = 709+499
                subtitle_bottomright_y = 622+60
        
        ## Define a basic funtion for calculating whether fixations in aoi or not.
        def isinaoi(fixation_x, fixation_y):            
            if ((fixation_x >= face_topleft_x and fixation_x <= face_bottomright_x) and 
               (fixation_y >= face_topleft_y and fixation_y <= face_bottomright_y)):
                return "face"
            elif ((fixation_x >= subtitle_topleft_x and fixation_x <= subtitle_bottomright_x) and 
               (fixation_y >= subtitle_topleft_y and fixation_y <= subtitle_bottomright_y)):
                return "subtitle"
            elif ((fixation_x >= slide_topleft_x and fixation_x <= slide_bottomright_x) and 
               (fixation_y >= slide_topleft_y and fixation_y <= slide_bottomright_y)): 
                return "slide"
            else:
                return "out"
        
        ## Define local features
        
        # num of saccade jump from one area to another
        num_saccade_aoi_face_out2in = 0
        num_saccade_aoi_face_aoi2in = 0
        num_saccade_aoi_face_in2out = 0
        num_saccade_aoi_face_in2aoi = 0
        num_saccade_aoi_face_within = 0
        
        num_saccade_aoi_slide_out2in = 0
        num_saccade_aoi_slide_aoi2in = 0
        num_saccade_aoi_slide_in2out = 0
        num_saccade_aoi_slide_in2aoi = 0
        num_saccade_aoi_slide_within = 0
        
        num_saccade_aoi_subtitle_out2in = 0
        num_saccade_aoi_subtitle_aoi2in = 0
        num_saccade_aoi_subtitle_in2out = 0
        num_saccade_aoi_subtitle_in2aoi = 0
        num_saccade_aoi_subtitle_within = 0
        
        temp_aoi = "out"
        
        # numbers and durations of fixations in AOIs.
        list_duration_fixation_aoi_face = []
        list_duration_fixation_aoi_subtitle = []
        list_duration_fixation_aoi_slide = []
        # fixations out of AOIs
        list_duration_fixation_aoi_out = []
        
        for index, row in df_GazeData_Tobii_selected.iterrows():
            if np.isnan(row['FixationIndex']):
                continue
            
            if temp_fixationindex == 0:
                temp_fixationindex = row['FixationIndex']
                temp_timestamp = row['Timestamp_utc']
                temp_FixationPointX = row['FixationPointX (MCSpx)']
                temp_FixationPointY = row['FixationPointY (MCSpx)']
                
                list_fixationduration.append(row['GazeEventDuration'])
                list_saccadeangel.append(row['AbsoluteSaccadicDirection'])
                
                ## calculate local features
                current_aoi = isinaoi(row['FixationPointX (MCSpx)'], row['FixationPointY (MCSpx)'])
                if current_aoi == "face":
                    list_duration_fixation_aoi_face.append(row['GazeEventDuration'])
                elif current_aoi == "subtitle":
                    list_duration_fixation_aoi_subtitle.append(row['GazeEventDuration'])
                elif current_aoi == "slide":
                    list_duration_fixation_aoi_slide.append(row['GazeEventDuration'])
                else:
                    list_duration_fixation_aoi_out.append(row['GazeEventDuration'])                
                temp_aoi = current_aoi
            
            elif temp_fixationindex != row['FixationIndex']:
                
                # Global features
                temp_fixationindex = row['FixationIndex']
                list_fixationduration.append(row['GazeEventDuration'])
                list_saccadeangel.append(row['AbsoluteSaccadicDirection'])
                
                datetime_previous = datetime.datetime.strptime(temp_timestamp, "%Y-%m-%dT%H:%M:%S.%fZ")
                datetime_current = datetime.datetime.strptime(row['Timestamp_utc'], "%Y-%m-%dT%H:%M:%S.%fZ")
                saccadeduration = datetime_current - datetime_previous
                list_saccadeduration.append(float(saccadeduration.total_seconds() * 1000))
                
                FixationPointX_current = row['FixationPointX (MCSpx)']
                FixationPointY_current = row['FixationPointY (MCSpx)']
                saccadedistance = math.sqrt(math.pow((FixationPointX_current - temp_FixationPointX), 2) + 
                                            math.pow((FixationPointY_current - temp_FixationPointY), 2))
                list_saccadedistance.append(saccadedistance)
                
                temp_timestamp = row['Timestamp_utc']
                temp_FixationPointX = row['FixationPointX (MCSpx)']
                temp_FixationPointY = row['FixationPointY (MCSpx)']
                
                # Local features
                current_aoi = isinaoi(row['FixationPointX (MCSpx)'], row['FixationPointY (MCSpx)'])
                
                if current_aoi == "face":
                    
                    list_duration_fixation_aoi_face.append(row['GazeEventDuration'])
                    
                    if temp_aoi == "face":
                        num_saccade_aoi_face_within = num_saccade_aoi_face_within + 1
                    elif temp_aoi == "out":
                        num_saccade_aoi_face_out2in = num_saccade_aoi_face_out2in + 1
                    else:
                        num_saccade_aoi_face_aoi2in = num_saccade_aoi_face_aoi2in + 1
                        if temp_aoi == "slide":
                            num_saccade_aoi_slide_in2aoi = num_saccade_aoi_slide_in2aoi + 1
                        else:
                            num_saccade_aoi_subtitle_in2aoi = num_saccade_aoi_subtitle_in2aoi + 1
                        
                elif current_aoi == "subtitle":
                    
                    list_duration_fixation_aoi_subtitle.append(row['GazeEventDuration'])
                    
                    if temp_aoi == "subtitle":
                        num_saccade_aoi_subtitle_within = num_saccade_aoi_subtitle_within + 1
                    elif temp_aoi == "out":
                        num_saccade_aoi_subtitle_out2in = num_saccade_aoi_subtitle_out2in + 1
                    else:
                        num_saccade_aoi_subtitle_aoi2in = num_saccade_aoi_subtitle_aoi2in + 1
                        if temp_aoi == "face":
                            num_saccade_aoi_face_in2aoi = num_saccade_aoi_face_in2aoi + 1
                        else:
                            num_saccade_aoi_slide_in2aoi = num_saccade_aoi_slide_in2aoi + 1
                    
                elif current_aoi == "slide":
                    list_duration_fixation_aoi_slide.append(row['GazeEventDuration'])
                    
                    if temp_aoi == "slide":
                        num_saccade_aoi_slide_within = num_saccade_aoi_slide_within + 1
                    elif temp_aoi == "out":
                        num_saccade_aoi_slide_out2in = num_saccade_aoi_slide_out2in + 1
                    else:
                        num_saccade_aoi_slide_aoi2in = num_saccade_aoi_slide_aoi2in + 1
                        if temp_aoi == "face":
                            num_saccade_aoi_face_in2aoi = num_saccade_aoi_face_in2aoi + 1
                        else:
                            num_saccade_aoi_subtitle_in2aoi = num_saccade_aoi_subtitle_in2aoi + 1
                    
                else:
                    list_duration_fixation_aoi_out.append(row['GazeEventDuration'])
                    if temp_aoi == "slide":
                        num_saccade_aoi_slide_in2out = num_saccade_aoi_slide_in2out + 1
                    elif temp_aoi == "face":
                        num_saccade_aoi_face_in2out = num_saccade_aoi_face_in2out + 1
                    elif temp_aoi == "subtitle":
                        num_saccade_aoi_subtitle_in2out = num_saccade_aoi_subtitle_in2out + 1
                 
                temp_aoi = current_aoi
            
            else:
                temp_timestamp = row['Timestamp_utc']
    
        
        num_saccade_horizon = sum(1 for i in list_saccadeangel if ((i <= 30 and i >= -30) or (i >= 150 and i <= 210) or (i >= 330)))
#         print num_saccade_horizon
#         print len(list_fixationduration)
#         print len(list_saccadeduration)
#         print len(list_saccadedistance)
#         print len(list_saccadeangel)
#         print len(list_duration_fixation_aoi_face)
#         print len(list_duration_fixation_aoi_subtitle)
#         print len(list_duration_fixation_aoi_slide)
#         print len(list_duration_fixation_aoi_out)
        
        duration_fixation_aoi_face = 0
        duration_fixation_aoi_face_max = 0
        duration_fixation_aoi_subtitle = 0
        duration_fixation_aoi_subtitle_max = 0
        duration_fixation_aoi_slide = 0
        duration_fixation_aoi_slide_max = 0
        duration_fixation_aoi_out = 0
        duration_fixation_aoi_out_max = 0 
        
        if len(list_duration_fixation_aoi_face) != 0:
            duration_fixation_aoi_face = sum(list_duration_fixation_aoi_face)/sum(list_fixationduration)
            duration_fixation_aoi_face_max = np.max(list_duration_fixation_aoi_face)
        if len(list_duration_fixation_aoi_subtitle) != 0:
            duration_fixation_aoi_subtitle = sum(list_duration_fixation_aoi_subtitle)/sum(list_fixationduration)
            duration_fixation_aoi_subtitle_max = np.max(list_duration_fixation_aoi_subtitle)
        if len(list_duration_fixation_aoi_slide) != 0:
            duration_fixation_aoi_slide = sum(list_duration_fixation_aoi_slide)/sum(list_fixationduration)
            duration_fixation_aoi_slide_max = np.max(list_duration_fixation_aoi_slide)
        if len(list_duration_fixation_aoi_out) != 0:
            duration_fixation_aoi_out = sum(list_duration_fixation_aoi_out)/sum(list_fixationduration)
            duration_fixation_aoi_out_max = np.max(list_duration_fixation_aoi_out)
        
        ## Add features into df_features
        df_features = df_features.append({
                'id': id_str, 
                'starttime_iso': starttime_iso, 
                'endtime_iso': endtime_iso,
                'fixationduration_min': np.min(list_fixationduration),
                'fixationduration_max': np.max(list_fixationduration),
                'fixationduration_mean': np.mean(list_fixationduration),
                'fixationduration_median': np.median(list_fixationduration),
                'fixationduration_stddev': np.std(list_fixationduration),
                'fixationduration_range': np.max(list_fixationduration) - np.min(list_fixationduration),
                'fixationduration_kurtosis': kurtosis(list_fixationduration),
                'fixationduration_skew': skew(list_fixationduration),
                'saccadeduration_min': np.min(list_saccadeduration),
                'saccadeduration_max': np.max(list_saccadeduration),
                'saccadeduration_mean': np.mean(list_saccadeduration),
                'saccadeduration_median': np.median(list_saccadeduration),
                'saccadeduration_stddev': np.std(list_saccadeduration),
                'saccadeduration_range': np.max(list_saccadeduration) - np.min(list_saccadeduration),
                'saccadeduration_kurtosis': kurtosis(list_saccadeduration),
                'saccadeduration_skew': skew(list_saccadeduration),
                'saccadedistance_min': np.min(list_saccadedistance),
                'saccadedistance_max': np.max(list_saccadedistance),
                'saccadedistance_mean': np.mean(list_saccadedistance),
                'saccadedistance_median': np.median(list_saccadedistance),
                'saccadedistance_stddev': np.std(list_saccadedistance),
                'saccadedistance_range': np.max(list_saccadedistance) - np.min(list_saccadedistance),
                'saccadedistance_kurtosis': kurtosis(list_saccadedistance),
                'saccadedistance_skew': skew(list_saccadedistance),
                'saccadeangel_min': np.min(list_saccadeangel),
                'saccadeangel_max': np.max(list_saccadeangel),
                'saccadeangel_mean': np.mean(list_saccadeangel),
                'saccadeangel_median': np.median(list_saccadeangel),
                'saccadeangel_stddev': np.std(list_saccadeangel),
                'saccadeangel_range': np.max(list_saccadeangel) - np.min(list_saccadeangel),
                'saccadeangel_kurtosis': kurtosis(list_saccadeangel),
                'saccadeangel_skew': skew(list_saccadeangel),
                'saccade_num': len(list_saccadeduration),
                'saccade_horizonratio': float(num_saccade_horizon)/len(list_saccadeangel),
                'fixation_saccade_ratio': sum(list_fixationduration)/sum(list_saccadeduration),
                ## LOCAL FEATURES
                'num_saccade_aoi_face_out2in': float(num_saccade_aoi_face_out2in)/len(list_saccadeduration),
                'num_saccade_aoi_face_aoi2in': float(num_saccade_aoi_face_aoi2in)/len(list_saccadeduration),
                'num_saccade_aoi_face_in2out': float(num_saccade_aoi_face_in2out)/len(list_saccadeduration),
                'num_saccade_aoi_face_in2aoi': float(num_saccade_aoi_face_in2aoi)/len(list_saccadeduration),
                'num_saccade_aoi_face_within': float(num_saccade_aoi_face_within)/len(list_saccadeduration),
                'num_saccade_aoi_slide_out2in': float(num_saccade_aoi_slide_out2in)/len(list_saccadeduration),
                'num_saccade_aoi_slide_aoi2in': float(num_saccade_aoi_slide_aoi2in)/len(list_saccadeduration),
                'num_saccade_aoi_slide_in2out': float(num_saccade_aoi_slide_in2out)/len(list_saccadeduration),
                'num_saccade_aoi_slide_in2aoi': float(num_saccade_aoi_slide_in2aoi)/len(list_saccadeduration),
                'num_saccade_aoi_slide_within': float(num_saccade_aoi_slide_within)/len(list_saccadeduration),
                'num_saccade_aoi_subtitle_out2in': float(num_saccade_aoi_subtitle_out2in)/len(list_saccadeduration),
                'num_saccade_aoi_subtitle_aoi2in': float(num_saccade_aoi_subtitle_aoi2in)/len(list_saccadeduration),
                'num_saccade_aoi_subtitle_in2out': float(num_saccade_aoi_subtitle_in2out)/len(list_saccadeduration),
                'num_saccade_aoi_subtitle_in2aoi': float(num_saccade_aoi_subtitle_in2aoi)/len(list_saccadeduration),
                'num_saccade_aoi_subtitle_within': float(num_saccade_aoi_subtitle_within)/len(list_saccadeduration),
                'duration_fixation_aoi_face': duration_fixation_aoi_face,
                'duration_fixation_aoi_face_max': duration_fixation_aoi_face_max,
                'duration_fixation_aoi_subtitle': duration_fixation_aoi_subtitle,
                'duration_fixation_aoi_subtitle_max': duration_fixation_aoi_subtitle_max,
                'duration_fixation_aoi_slide': duration_fixation_aoi_slide,
                'duration_fixation_aoi_slide_max': duration_fixation_aoi_slide_max,
                'duration_fixation_aoi_out': duration_fixation_aoi_out,
                'duration_fixation_aoi_out_max': duration_fixation_aoi_out_max
            }, ignore_index=True)
        
        # print df_features.head(1)
print df_features.shape

## Merge features with labels
df_merge = pd.merge(df_reports, df_features)
# df_merge.to_csv("features_labels_webgazer.csv", index=False)
df_merge.to_csv("features_labels_tobii_sd.csv", index=False)

print df_merge.columns
print df_merge.columns[10:18]
print df_merge.columns[18:27]
print df_merge.columns[27:42]
print df_merge.columns[42:]
print len(df_merge.columns)

## Generate the results of the baseline
Assumin that we know the mind-wandering rate is 0.29, the baseline is the precision, recall and f1 when we totally guess the result with 0.29 mind-wandering rate.

from random import random, seed
from sklearn.metrics import (precision_score, recall_score, f1_score, accuracy_score)

# read the data
df_merge = pd.read_csv("features_labels_webgazer.csv")
# read the label
y_test = df_merge.ix[:, 4]
print df_merge.shape

repeat_parameter = 10000
seed_parameter = 48 # same as the example in SMOTE

list_test = list()
for i in range(0, repeat_parameter):
    list_test = list_test + list(y_test)

seed(seed_parameter)
list_guess = list()
for i in range(0, len(list_test)):
    temp_random = random()
    if temp_random > 0.29:
        list_guess.append(0)
    else:
        list_guess.append(1)

print "Baseline with 0.29 mind-wandering rate"
print("\tPrecision: %1.3f" % precision_score(list_test, list_guess))
print("\tRecall: %1.3f" % recall_score(list_test, list_guess))
print("\tF1: %1.3f" % f1_score(list_test, list_guess))
print("\tAccuracy: %1.3f\n" % accuracy_score(list_test, list_guess))

## Prediction With Different Features and Models

Nested cross validation is used in our experiment.
Each time we select data of one participant as test data and the remaining as training data. Then, for the training data we use leave-one-paticipant-out cross validation to selection model and predict the result of the test data.

import pandas as pd
import numpy as np
from pandas import DataFrame

from imblearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.svm import LinearSVC
from sklearn.naive_bayes import GaussianNB
from imblearn.over_sampling import SMOTE
from sklearn.model_selection import (GridSearchCV, LeaveOneGroupOut)
from sklearn.metrics import (precision_score, recall_score, f1_score, accuracy_score)

# df_merge = pd.read_csv("features_labels_webgazer.csv")
df_merge = pd.read_csv("features_labels_tobii_sd.csv")

# Remove 3 participants since they only have 1 mind-wandering report in their test
df_merge = df_merge[df_merge.id != "Anon06"]
df_merge = df_merge[df_merge.id != "Anon07"]
df_merge = df_merge[df_merge.id != "Anon13"]

list_id_report = list(df_merge['id'].unique())

# Define different features running in the experiments
feature_index_local = [10,11,12,13,14,15,16,17,
                       27,28,29,30,31,32,33,34,35,36,37,38,39,40,41]
feature_index_local_dict = {"feature_name": "Local Features", 
                            "feature_index": feature_index_local}
feature_index_global = [18,19,20,21,22,23,24,25,26,
                        42,43,44,45,46,47,48,49,50,
                        51,52,53,54,55,56,57,58,59,
                        60,61,62,63,64,65,66,67]
feature_index_global_dict = {"feature_name": "Global Features", 
                             "feature_index": feature_index_global}
feature_index_all = range(10, 68)
feature_index_all_dict = {"feature_name": "All Features", 
                          "feature_index": feature_index_all}
featuren_list = [feature_index_all_dict, 
                 feature_index_global_dict, 
                 feature_index_local_dict]

# Define pipelines
sm = SMOTE(random_state=48)

pipe_svm_dict = {"pipe_name": "Linear SVM without SMOTE methods",
                 "pipe": Pipeline([('classify', LinearSVC())]),
                 "p_grid": {"classify__loss": ['hinge', 'squared_hinge'],
                            "classify__tol": [0.1, 0.01, 0.001, 0.0001],
                            "classify__C": [100, 10, 1, 0.1],
                            "classify__class_weight": [None, 'balanced']}
                }
pipe_svm_sm_dict = {"pipe_name": "Linear SVM with SMOTE methods",
                    "pipe": Pipeline([('balanced', sm), 
                                      ('classify', LinearSVC())]),
                    "p_grid": {"classify__loss": ['hinge', 'squared_hinge'],
                               "classify__tol": [0.1, 0.01, 0.001, 0.0001],
                               "classify__C": [100, 10, 1, 0.1],
                               "classify__class_weight": [None, 'balanced']}}
pipe_lr_dict = {"pipe_name": "Logistic Regression without SMOTE methods", 
                "pipe": Pipeline([('classify', LogisticRegression())]),
                "p_grid": {"classify__penalty": ['l1', 'l2'],
                           "classify__tol": [0.1, 0.01, 0.001, 0.0001],
                           "classify__C": [100, 10, 1, 0.1],
                           "classify__class_weight": [None, 'balanced']}
               }
pipe_lr_sm_dict = {"pipe_name": "Logistic Regression with SMOTE methods", 
                   "pipe": Pipeline([('balanced', sm), 
                                     ('classify', LogisticRegression())]),
                   "p_grid": {"classify__penalty": ['l1', 'l2'],
                              "classify__tol": [0.1, 0.01, 0.001, 0.0001],
                              "classify__C": [100, 10, 1, 0.1],
                              "classify__class_weight": [None, 'balanced']}
                  }
pipe_nb_dict = {"pipe_name": "Naive Bayes without SMOTE methods",
                "pipe": Pipeline([('classify', GaussianNB())]),
                "p_grid": {}
               }
pipe_nb_sm_dict = {"pipe_name": "Naive Bayes with SMOTE methods",
                   "pipe": Pipeline([('balanced', sm), 
                                     ('classify', GaussianNB())]),
                   "p_grid": {}
                  }

pipe_list = [pipe_svm_dict, 
             pipe_svm_sm_dict, 
             pipe_lr_dict, 
             pipe_lr_sm_dict, 
             pipe_nb_dict, 
             pipe_nb_sm_dict]

for feature_index_dict in featuren_list:
    featurn_name = feature_index_dict["feature_name"]
    feature_index = feature_index_dict["feature_index"]
    
    print "-----------------------------------"
    print featurn_name
    print "-----------------------------------"
    
    for pipe_dict in pipe_list:
        print pipe_dict["pipe_name"]

        y_test_total = []
        y_pred_total = []
        f1_list = []
        precision_list = []
        recall_list = []

        ## Leave-one-participant-out machine learning methods (try Logistic Regression first)
        for i in range(0, len(list_id_report)):

            id_str = list_id_report[i]
            print id_str

            # outer cv, split data for the training and test
            data_train = df_merge.ix[df_merge['id'] != id_str]
            data_test = df_merge.ix[df_merge['id'] == id_str]

            # inner cv, data preparation
            X_train = data_train.ix[:, feature_index].fillna(value=0)
            y_train = list(data_train.ix[:, 4])
            group = list(data_train.ix[:, 3])

            # inner cv, we define the pipeline 
            pipe = pipe_dict["pipe"]
            p_grid = pipe_dict["p_grid"]
            inner_cv = LeaveOneGroupOut()

            # inner cv, model selection by gridsearch
            clf = GridSearchCV(estimator=pipe, param_grid=p_grid, cv=inner_cv, scoring='f1')
            clf.fit(X_train, y_train, group)
            print "inner cv parameters: " + str(clf.best_params_)
            print "inner cv f1 score: " + str(clf.best_score_)

            # outer cv, data preparation
            X_test = data_test.ix[:, feature_index].fillna(value=0)
            y_test = data_test.ix[:, 4]

            y_test_total.extend(y_test)

            y_pred = clf.predict(X_test)
            y_pred_total.extend(y_pred)

            print("outter cv results")
            print("\tPrecision: %1.3f" % precision_score(y_test, y_pred))
            print("\tRecall: %1.3f" % recall_score(y_test, y_pred))
            print("\tF1: %1.3f" % f1_score(y_test, y_pred))
            print("\tAccuracy: %1.3f\n" % accuracy_score(y_test, y_pred))
            
            f1_list.append(f1_score(y_test, y_pred))
            precision_list.append(precision_score(y_test, y_pred))
            recall_list.append(recall_score(y_test, y_pred))

        print "Nested cv results"
#         print("\tPrecision: %1.3f" % precision_score(y_test_total, y_pred_total))
#         print("\tRecall: %1.3f" % recall_score(y_test_total, y_pred_total))
#         print("\tF1: %1.3f" % f1_score(y_test_total, y_pred_total))
#         print("\tAccuracy: %1.3f\n" % accuracy_score(y_test_total, y_pred_total))
        print("\tF1: %1.3f" % np.mean(f1_list))
        print("\tPrecision: %1.3f" % np.mean(precision_list))
        print("\tRecall: %1.3f" % np.mean(recall_list))

## 1.1 Whether or not predictions can be made equally well for all participants?
Check results on individual in above part.

## 1.2 whether or not predictions can be made well across the entire length of the lecture videos?

import numpy as np
import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.naive_bayes import GaussianNB
from imblearn.over_sampling import SMOTE

from sklearn.metrics import (precision_score, recall_score, f1_score, accuracy_score)
## Read data from features_labels_tobii.csv

df_merge = pd.read_csv("features_labels_webgazer.csv")
list_id_report = list(df_merge['id'].unique())

y_test_total = []
y_pred_bayes_total = []

list_pre = []
list_rec = []
list_f1 = []

## Leave-one-out machine learning methods (try Logistic Regression first)
for i in range(0, len(list_id_report)):
    id_str = list_id_report[i]    
    
    data_train = df_merge
    data_train = data_train.loc[((data_train['video_id'] != "Solar")&
                                (data_train['endtime_video']/data_train['video_length'] < 0.5))|
                               (data_train['video_id'] == "Solar")
                              ]
    
    # print data_train.shape
    data_test = df_merge.loc[df_merge['id'] == id_str]
    data_test = data_test.loc[(data_test['video_id'] != "Solar")&(data_test['endtime_video']/data_test['video_length'] >= 0.5)]
    # print data_test.shape
    # print data_test.shape
    feature_index_global = [18,19,20,21,22,23,24,25,26,
                        42,43,44,45,46,47,48,49,50,
                        51,52,53,54,55,56,57,58,59,
                        60,61,62,63,64,65,66,67]
    X_train = data_train.iloc[:, feature_index_global].fillna(value=0)
    y_train = data_train.iloc[:, 4]
    
    sm = SMOTE(random_state=48)
    X_train, y_train = sm.fit_sample(X_train, y_train)
    
    X_test = data_test.iloc[:, feature_index_global].fillna(value=0)
    y_test = data_test.iloc[:, 4]
    
    y_test_total.extend(y_test)
    
    ## Logistic Regression
    
    clf_GaussianNB = GaussianNB()
    clf_GaussianNB = clf_GaussianNB.fit(X_train, y_train)
    y_pred_bayes = clf_GaussianNB.predict(X_test)
    y_pred_bayes_total.extend(y_pred_bayes)
    
    list_pre.append(precision_score(y_test, y_pred_bayes))
    list_rec.append(recall_score(y_test, y_pred_bayes))
    list_f1.append(f1_score(y_test, y_pred_bayes))

print "Final Results: GaussianNB"
print("\tPrecision: %1.3f" % np.mean(list_pre))
print("\tRecall: %1.3f" % np.mean(list_rec))
print("\tF1: %1.3f" % np.mean(list_f1))

## 1.3 whether or not a model trained on one video translates to good predictions in other videos?

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.naive_bayes import GaussianNB
from imblearn.over_sampling import SMOTE

from sklearn.metrics import (precision_score, recall_score, f1_score, accuracy_score)
## Read data from features_labels_tobii.csv

# df_merge = pd.read_csv("features_labels_tobii.csv")
df_merge = pd.read_csv("features_labels_webgazer.csv")
list_id_report = list(df_merge['id'].unique())

y_test_total = []
y_pred_bayes_total = []

list_pre = []
list_rec = []
list_f1 = []

## Leave-one-out machine learning methods (try Logistic Regression first)
for i in range(0, len(list_id_report)):
    id_str = list_id_report[i]    
    
    #data_train = df_merge.loc[df_merge['id'] != id_str]
    data_train = data_train.loc[data_train['video_id'] != "Solar"]
    
    # print data_train.shape
    data_test = df_merge.loc[df_merge['id'] == id_str]
    data_test = data_test.loc[data_test['video_id'] == "Solar"]
    
    feature_index_global = [18,19,20,21,22,23,24,25,26,
                        42,43,44,45,46,47,48,49,50,
                        51,52,53,54,55,56,57,58,59,
                        60,61,62,63,64,65,66,67]
    X_train = data_train.iloc[:, feature_index_global].fillna(value=0)
    y_train = data_train.iloc[:, 4]
    
    sm = SMOTE(random_state=48)
    X_train, y_train = sm.fit_sample(X_train, y_train)
    
    X_test = data_test.iloc[:, feature_index_global].fillna(value=0)
    y_test = data_test.iloc[:, 4]
    
    y_test_total.extend(y_test)
    
    ## Logistic Regression
    
    clf_GaussianNB = GaussianNB()
    clf_GaussianNB = clf_GaussianNB.fit(X_train, y_train)
    y_pred_bayes = clf_GaussianNB.predict(X_test)
    y_pred_bayes_total.extend(y_pred_bayes)
    
    list_pre.append(precision_score(y_test, y_pred_bayes))
    list_rec.append(recall_score(y_test, y_pred_bayes))
    list_f1.append(f1_score(y_test, y_pred_bayes))

print "Final Results: GaussianNB"
print("\tPrecision: %1.3f" % np.mean(list_pre))
print("\tRecall: %1.3f" % np.mean(list_rec))
print("\tF1: %1.3f" % np.mean(list_f1))
